Tests for ProductController and CartController.

+++ src/main/java/com/uedsonreis/ecommerce/services/ProductService.java
@@ -27,6 +27,7 @@ public class ProductService {
 	}
 	
 	public Boolean remove(int id) {
+		if (id < 0 || id >= this.repository.size()) return false;
 		return (this.repository.remove(id) != null);
 	}
